Move full-text indexing tests to their own test project
