+++ Libraries/dotNetRDF.Query.FullText/dotNetRDF.Query.FullText.csproj
@@ -24,7 +24,7 @@
     <ProjectReference Include="..\dotNetRdf.InMemory\dotNetRdf.InMemory.csproj" />
     <ProjectReference Include="..\dotNetRdf.Sparql\dotNetRdf.Sparql.csproj" />
     <ProjectReference Include="..\dotNetRDF\dotNetRDF.csproj" />
-    <InternalsVisibleTo Include="dotNetRdf.Test" />
+    <InternalsVisibleTo Include="dotNetRdf.Query.FullText.Tests" />
   </ItemGroup>
 
   <ItemGroup>
